1. 根据数据库里配置的job信息，实现启动时自动运行job。 2. 未完成：针对job配置的管理，以及job的代理IP抓取逻辑处理，

+++ admin/src/main/java/com/cv4j/netdiscovery/admin/dto/PageResult.java
@@ -16,7 +16,7 @@ public class PageResult<T> {
 	
 	private String msg = "";  //提示信息
 
-	private int count; // 总数量,
+	private long count; // 总数量,
 
 	private List<T> data; // 当前页数据
 
@@ -24,7 +24,7 @@ public class PageResult<T> {
 		this.code = code;
 	}
 
-	public PageResult(List<T> data, int count) {
+	public PageResult(List<T> data, long count) {
 		this.data = data;
 		this.count = count;
 	}
